Renamed Queue to submitCommands
