+++ tests/utils/test_StackArray.cpp
@@ -74,8 +74,8 @@ UTEST(test_MSArray, CreateMSArrayWithSmallerInitializerList)
 
     ASSERT_EQ(2, array.Count());
 
-    ASSERT_FALSE(array.Active(2));
-    ASSERT_FALSE(array.Active(3));
+    ASSERT_FALSE(array.IsActive(2));
+    ASSERT_FALSE(array.IsActive(3));
 }
 
 UTEST(test_MSArray, CreateMSArrayFromCopyConstructor)
@@ -135,10 +135,10 @@ UTEST(test_MSArray, CreateMSArrayFromRawArray)
     ASSERT_EQ(array.Size(), 4);
     ASSERT_NE(array.Data(), nullptr);
 
-    ASSERT_TRUE(array.Active(0));
-    ASSERT_TRUE(array.Active(1));
-    ASSERT_TRUE(array.Active(2));
-    ASSERT_TRUE(array.Active(3));
+    ASSERT_TRUE(array.IsActive(0));
+    ASSERT_TRUE(array.IsActive(1));
+    ASSERT_TRUE(array.IsActive(2));
+    ASSERT_TRUE(array.IsActive(3));
 
     ASSERT_EQ(array[0], 1);
     ASSERT_EQ(array[1], 2);
@@ -155,10 +155,10 @@ UTEST(test_MSArray, CreateMSArrayFromLargerRawArray)
     ASSERT_EQ(array.Size(), 4);
     ASSERT_NE(array.Data(), nullptr);
 
-    ASSERT_TRUE(array.Active(0));
-    ASSERT_TRUE(array.Active(1));
-    ASSERT_TRUE(array.Active(2));
-    ASSERT_TRUE(array.Active(3));
+    ASSERT_TRUE(array.IsActive(0));
+    ASSERT_TRUE(array.IsActive(1));
+    ASSERT_TRUE(array.IsActive(2));
+    ASSERT_TRUE(array.IsActive(3));
 
     ASSERT_EQ(array[0], 1);
     ASSERT_EQ(array[1], 2);
@@ -175,10 +175,10 @@ UTEST(test_MSArray, CreateMSArrayFromSmallerRawArray)
     ASSERT_EQ(array.Size(), 4);
     ASSERT_NE(array.Data(), nullptr);
 
-    ASSERT_TRUE(array.Active(0));
-    ASSERT_TRUE(array.Active(1));
-    ASSERT_FALSE(array.Active(2));
-    ASSERT_FALSE(array.Active(3));
+    ASSERT_TRUE(array.IsActive(0));
+    ASSERT_TRUE(array.IsActive(1));
+    ASSERT_FALSE(array.IsActive(2));
+    ASSERT_FALSE(array.IsActive(3));
 
     ASSERT_EQ(array[0], 1);
     ASSERT_EQ(array[1], 2);
@@ -266,13 +266,13 @@ UTEST(test_MSArray, AppendValuesToArray)
     array.Append(1);
 
     ASSERT_EQ(1, array.Count());
-    ASSERT_TRUE(array.Active(0));
+    ASSERT_TRUE(array.IsActive(0));
     ASSERT_EQ(2, array.Size());
 
     array.Append(2);
 
     ASSERT_EQ(2, array.Count());
-    ASSERT_TRUE(array.Active(1));
+    ASSERT_TRUE(array.IsActive(1));
     ASSERT_EQ(2, array.Size());
 
     ASSERT_EQ(1, array[0]);
@@ -311,12 +311,12 @@ UTEST(test_MSArray, CheckElementIsActive)
 
     array[0] = 1;
 
-    ASSERT_TRUE(array.Active(0));
-    ASSERT_FALSE(array.Active(1));
+    ASSERT_TRUE(array.IsActive(0));
+    ASSERT_FALSE(array.IsActive(1));
 
     array[1] = 2;
 
-    ASSERT_TRUE(array.Active(1));
+    ASSERT_TRUE(array.IsActive(1));
 }
 
 UTEST(test_MSArray, RemoveElementByIndex)
@@ -326,22 +326,22 @@ UTEST(test_MSArray, RemoveElementByIndex)
     array[0] = 1;
 
     ASSERT_EQ(array.Count(), 1);
-    ASSERT_TRUE(array.Active(0));
+    ASSERT_TRUE(array.IsActive(0));
 
     array[1] = 2;
 
     ASSERT_EQ(array.Count(), 2);
-    ASSERT_TRUE(array.Active(1));
+    ASSERT_TRUE(array.IsActive(1));
 
     array.Remove(1);
 
     ASSERT_EQ(array.Count(), 1);
-    ASSERT_FALSE(array.Active(1));
+    ASSERT_FALSE(array.IsActive(1));
 
     array.Remove(0);
 
     ASSERT_EQ(array.Count(), 0);
-    ASSERT_FALSE(array.Active(0));
+    ASSERT_FALSE(array.IsActive(0));
     ASSERT_TRUE(array.Empty());
 }
